+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 # Project Manager Changelog
 
-## [Initial Version] - {PR_MERGE_DATE}
+## [Initial Version] - 2025-03-20
+
+- Initial version
+
+## [1.0.1] - 2025-03-22
+
+- Add copy project name and path actions
+- Add code editor and terminal picker, instead of textfield
+- Fix bug when open project in iTerm